requirements.txt
fastapi
uvicorn
python-multipart
requests
twilio
SQLAlchemy
python-dotenv
